requirements.txt
pandas
matplotlib
seaborn
pillow
torch
torchvision
timm
scikit-learn
joblib
albumentations
